added matches component
